C programming

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -13,10 +13,10 @@ int print_last_digit(int n)
 {
 	if (n >= 0)
 	{
-		return (int(n%10));
+		return (n%10);
 	}
 	else
 	{
-                return (int((-1*n)%10));
+                return ((-1*n)%10);
 	}
 }
